Added KeysRotator. Added more logs
